backend/requirements.txt
fastapi==0.111.1
openai==1.37.0
pandas==2.2.2
openpyxl==3.1.5
uvicorn==0.30.6
asyncpg==0.29.0 
langchain_pinecone==0.1.3
pinecone-client==5.0.1
langchain_openai==0.1.20
Markdown==3.7
beautifulsoup4==4.12.3